backend/requirements.txt
Flask
flask-cors
google-cloud-storage
google-cloud-vision
torch
torchvision
opencv-python-headless
numpy==1.26.3
pandas
flask_cors
flask
gunicorn
waitress==2.1.2
asgiref==3.5.2
firebase_admin